feat: adds basic lint command and initial schema
